Add a server shutdown registration to shrinksafe
Shrinksafe has an option for starting child processes in a server mode
to allow them to be used more than once.  There had been no way to
explicitly shut them down; there hadn't been a reason to, relying on the
shutdown of the parent javascript engine process to get the children
shutdown by the OS.

But the node parent process won't exit while it has event handlers
registered for events that might still happen, and having child
processes running with their own stdin/stdout/stderr tied to node
streams is reason for there to be event handlers registered for events
that might still fire.

So the narwhal process would hang, waiting for the event handles to be
removed, or the streams to be closed by the OS.  The java server process
that had been started would not see its stdin close, so would not
shutdown, and therefore the stdout and stderr streams remained open, and
their handlers remaining registered.

With this addition, shrinksafe provides a callback that gets the stdin,
and stdout and stderr closed.  Closing stdin causes the java server
process to shutdown gracefully, also causing the stdout and stderr to be
closed.

If it becomes necessary (for other process reasons), the narwhal wrapper
for node could go to extra lengths to track all event handlers that it
creates for the streams created by popen calls, and then remove them
itself when it is expecting to wrap up execution.  But this current
approach seems straight forward and reasonable: shrinksafe knows if it
has started servers and it can provide a callback to get the servers
shutdown at the end.

For those javascript engines that don't care, they don't provide the
callback registration mechanism.  So there is no change for jsc and
rhino, how their children server processes are shutdown is unaffected.