Add node js as a NARWHAL_ENGINE
This adds a new NARWHAL_ENGINE based on node.js.

Tested on Darwin and Linux.

After installing node and npm, and after running the
narwhal/packages/narwhal-node/bin/npm-install script to get the required
npm packages installed,

This new package follows closely the API that the jsc narwhal-engine
implemented.  There were a lot of TBDs and half started projects in
the cappuccino-base code so it was hard to figure out what was needed
and what was not.  More may be supported than is necessary but also
something might still be missing that testing hasn't brought to light
yet.

Here node is wrapped by js code defined as the narwhal-node package, as
opposed to the narwhal-jsc package - bridging the gap between the
narwhal API and the node API.  The narwhal-node code freely uses ES6
syntax and has been tested on Darwin and Linux.

Node is generally used in an asynchronous manner, employing callbacks for
operating system access.  But asynchronous behaviour, if exposed, would
not be compatible with the function of narwhal, namely to be the engine
for Jake and the Objective-J compiler.  Both Jake and the Objective-J
compiler are written for a synchronous engine, where calls block until
they are complete, and there is a single call stack.

Node provides some synchronous functions now but some node functionality
is needed to support narwhal that still requires asynchronous access.
Fortunately, others have also wanted to use node in a synchronous manner
so an npm package named fibers exists and is used here.  Fibers
implements additional call stacks within node and lets the user code
wait on its own stack, suspended, while the node event queue is being
processed, and when the event that the user code wanted to wait for has
completed, the user code's stack is resumed.  So the narwhal-node
package presents a synchronous usage model to the higher level codes.

This has been tested with the unittests that are part of cappuccino-base
and with the cappuccino repository.  One change (could be viewed as a
small fix actually) was needed to the cappuccino code base to be
compatible with this commit.  So if this is being tried, pull the
cappuccino fork from my repository or cherry pick the commit that was
made there.  Additional cleanup to the unittests will be made as
subsequent commits, to let folks cherry pick the commits they are
interested in trying out.

To run:

    - install cappuccino-base and cappuccino from my forks
    - install node.js and npm.
    - read and run ~/narwhal/packages/narwhal-node/bin/npm-install
    - set environment variable NARWHAL_ENGINE=node
    - call 'jake build' or 'jake install' from ~/cappuccino