Fix a forEach test so node can pass it.
Here is a test that jsc and rhino had been passing but it is being
simplified so that node can pass it too.

Replaced with a forEach test that does not rely on calling the forEach
twice.  This is functionality that jsc and rhino support but that is not
used by the cappuccino framework for building cappuccino code.

Not sure what it means to loop while there is a line of text to read
without blocking until the stream is either closed or you found what you
were waiting for.

Also the idea of wanting to read an input stream when the output stream
that is feeding it hasn't been closed first doesn't seem strong or
important for cappuccino.  Cappuccino can require streams be closed
before the file be read by something else.