Base the node engine's dom support on xmldom
xmldom is an npm package that gets installed along with the fibers and
synchronize packages by the
narwhal/packages/narwhal-node/bin/npm-install script.  The xmldom
package is a single package with no dependencies on other packages.

xmldom provides DOMParser() and XMLSerializer().  Both are exported by
the node engine's dom.js module.

The xmldom author provides for either of two LICENSES to be used.  We
will choose the MIT license.